+++ package.json
@@ -7,6 +7,7 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "firebase": "^9.17.1",
+    "nodemailer": "^6.9.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.7.1",
@@ -40,6 +41,7 @@
     ]
   },
   "devDependencies": {
+    "@types/nodemailer": "^6.4.7",
     "react-router-dom": "^6.8.1"
   }
 }
