+++ src/components/AvatarChangeBox/AvatarChangeBox.jsx
@@ -6,9 +6,13 @@ import { MdFileUpload } from 'react-icons/md';
 import { MdCancel } from 'react-icons/md';
 
 const PromptingCloud = ({ handleClose, handleAvatarChange, handleUpload }) => {
+
+
     return (
         <div className="promptingAvatarChangeCloud">
-            <h2>Upload New Avatar</h2>
+            <h3>Upload A New Avatar</h3>
+            <span>Press on the image icon to add an Avatar</span>
+            <br />
             <label>
                 <BiImageAdd className="imageInput" />
                 {/* to customize standard input look we hide the input element and wrap it in a
