+++ src/pages/not-found/NotFound.scss
@@ -1,6 +1,6 @@
 @import '../../index.scss';
 
-.mainbox {
+.notFoundContainer {
     background-color: $mainLight;
     margin: auto;
     height: 100vh;
